Outsource Column and Particle data classes
