+++ Equation-Generator.py
@@ -24,747 +24,7 @@ from src import load_CADET_h5
 from src import ui_config
 from src.utils import format_variables
 from src.renderer import availability_badge_html, write_and_save as renderer_write_and_save
-
-
-# immutable & hashable dataclass -> unique particle type counter
-@dataclass(frozen=True)
-class Particle:
-
-    geometry: Literal["Sphere", "Cylinder", "Slab"]
-    has_core: bool
-    resolution: Literal["1D", "0D"]
-    has_binding: bool =None # and thus solid phase
-    req_binding: bool = None
-    has_mult_bnd_states: bool = None
-    has_surfDiff: bool = None
-    nonlimiting_filmDiff: bool = None
-    surface_volume_ratio: float = None
-    interstitial_volume_resolution: str = None
-    single_partype: bool = True
-    PTD: bool = False
-    binding_model: str = "Arbitrary"
-    has_reaction_liquid: bool = False
-    has_reaction_solid: bool = False
-    # volume fraction ?
-    # binding -> is_kinetic, nBound
-    vars_and_params: List[dict] = field(default_factory=list, init=False, compare=False, hash=False)
-
-    def __post_init__(self):
-        valid_geometries = {"Sphere", "Cylinder", "Slab"}
-        valid_resolutions = {"1D", "0D"}
-
-        if self.geometry not in valid_geometries:
-            raise ValueError(
-                f"Invalid geometry: {self.geometry}. Must be one of {valid_geometries}.")
-        if self.resolution not in valid_resolutions:
-            raise ValueError(
-                f"Invalid resolution: {self.resolution}. Must be one of {valid_resolutions}.")
-
-        if self.geometry == "Sphere":
-            object.__setattr__(self, 'surface_volume_ratio', 3)
-        elif self.geometry == "Cylinder":
-            object.__setattr__(self, 'surface_volume_ratio', 2)
-        elif self.geometry == "Slab":
-            object.__setattr__(self, 'surface_volume_ratio', 1)
-
-        vars_and_params_ = []
-        
-        state_deps = r"t"
-        if self.interstitial_volume_resolution == "1D":
-            state_deps = r"t, z"
-        if self.interstitial_volume_resolution == "2D":
-            state_deps = r"t, z, \rho"
-        if self.interstitial_volume_resolution == "3D":
-            state_deps = r"t, z, \rho, \phi"
-        if self.resolution == "1D":
-            state_deps += r", r"
-
-        state_deps += "; i" if self.single_partype else "; j, i"
-
-        if not (self.nonlimiting_filmDiff and self.resolution == "0D"):
-            symbol_name_ = r"c^{\p}_{i}" if self.single_partype else r"c^{\p}_{j,i}"
-            vars_and_params_.append({"Group" : 1, "Symbol": symbol_name_, "Description": r"particle liquid concentration", "Unit": r"\frac{mol}{m^3}", "Dependence" : state_deps, "Domain" : eq.full_particle_conc_domain(column_resolution=self.interstitial_volume_resolution, particle_resolution=self.resolution, hasCore=self.has_core, with_par_index=False, with_time_domain=True)})
-            
-        if self.resolution == "1D":
-            vars_and_params_.append({"Group" : 0, "Symbol": r"r", "Description": r"radial particle coordinate", "Unit": r"m", "Dependence": r"\text{independent variable}", "Property": r""})
-            symbol_name_ = r"D^\mathrm{p}_{i}" if self.single_partype else r"D^\mathrm{p}_{j,i}"
-            vars_and_params_.append({"Group" : 6.1, "Symbol": symbol_name_, "Description": r"particle diffusion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": r"i" if self.single_partype else r"j,i", "Property": r"> 0"})
-
-            if not self.nonlimiting_filmDiff:
-                vars_and_params_.append({"Group" : 4, "Symbol": r"\varepsilon^{\mathrm{p}}" if self.single_partype else r"\varepsilon^{\mathrm{p}}_{j}", "Description": r"particle porosity", "Unit": r"-", "Dependence": r"\text{constant}" if self.single_partype else r"j", "Property": r"\in (0, 1)"})
-
-        if self.has_binding:
-            symbol_name_ = r"c^{\s}_{i}" if self.single_partype else r"c^{\s}_{j,i}"
-            vars_and_params_.append({"Group" : 1, "Symbol": symbol_name_, "Description": r"particle solid concentration", "Unit": r"\frac{mol}{m^3}", "Dependence" : state_deps, "Domain" : eq.full_particle_conc_domain(column_resolution=self.interstitial_volume_resolution, particle_resolution=self.resolution, hasCore=self.has_core, with_par_index=False, with_time_domain=True)})
-
-            if self.binding_model == "Arbitrary":
-                symbol_name_ = r"f^\mathrm{bind}_{j,i}" if self.PTD else r"f^\mathrm{bind}_{i}"
-                dep_ = r"\vec{c}^\mathrm{p}, \vec{c}^\mathrm{s}; j, i" if self.PTD else r"\vec{c}^\mathrm{p}, \vec{c}^\mathrm{s}; i"
-                vars_and_params_.append({"Group" : 10, "Symbol": symbol_name_, "Description": r"adsorption isotherm function", "Unit": r"\frac{1}{s}", "Dependence": dep_})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"\vec{c}^\mathrm{p}", "Description": r"particle liquid components vector", "Unit": r"[\frac{mol}{m^3}]", "Dependence": state_deps})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"\vec{c}^\mathrm{s}", "Description": r"particle solid components vector", "Unit": r"[\frac{mol}{m^3}]", "Dependence": state_deps})
-            else:
-                idx_ = "i" if self.single_partype else "j,i"
-                vars_and_params_.append({"Group" : 10, "Symbol": r"k^{\mathrm{a}}_{" + idx_ + r"}", "Description": r"adsorption rate constant", "Unit": r"\frac{m^3}{mol \cdot s}", "Dependence": idx_})
-                vars_and_params_.append({"Group" : 10, "Symbol": r"k^{\mathrm{d}}_{" + idx_ + r"}", "Description": r"desorption rate constant", "Unit": r"\frac{1}{s}", "Dependence": idx_})
-
-            if self.binding_model == "Langmuir":
-                idx_ = "i" if self.single_partype else "j,i"
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"q^{\mathrm{max}}_{" + idx_ + r"}", "Description": r"maximum binding capacity", "Unit": r"\frac{mol}{m^3}", "Dependence": idx_})
-
-            if self.binding_model == "SMA":
-                idx_ = "i" if self.single_partype else "j,i"
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"\nu_{" + idx_ + r"}", "Description": r"characteristic charge", "Unit": r"-", "Dependence": idx_})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"\sigma_{" + idx_ + r"}", "Description": r"steric factor", "Unit": r"-", "Dependence": idx_})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"\Lambda", "Description": r"ionic capacity (binding site concentration)", "Unit": r"\frac{mol}{m^3}", "Dependence": r"\text{constant}"})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"q^{\mathrm{ref}}", "Description": r"reference solid phase concentration", "Unit": r"\frac{mol}{m^3}", "Dependence": r"\text{constant}"})
-                vars_and_params_.append({"Group" : 10.1, "Symbol": r"c^{\mathrm{ref}}", "Description": r"reference liquid phase concentration", "Unit": r"\frac{mol}{m^3}", "Dependence": r"\text{constant}"})
-            
-            if self.has_surfDiff:
-                symbol_name_ = r"D^\mathrm{s}_{i}" if self.single_partype else r"D^\mathrm{s}_{j,i}"
-                vars_and_params_.append({"Group" : 6.1, "Symbol": symbol_name_, "Description": r"surface diffusion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": r"i" if self.single_partype else r"j,i", "Property": r"\geq 0"})
-            
-            if self.has_mult_bnd_states:
-                symbol_name_ = r"N^{\mathrm{b}}_{i}" if self.single_partype else r"N^{\mathrm{b}}_{j,i}"
-                vars_and_params_.append({"Group" : 11, "Symbol": symbol_name_, "Description": r"number of bound states", "Unit": r"-", "Dependence": r"i" if self.single_partype else r"j,i"})
-            
-            if self.has_core:
-                self.vars_and_params.append({"Group" : 0.1, "Symbol": r"R^\mathrm{pc}", "Description": r"particle core radius", "Unit": r"-", "Dependence": r"-", "Property": r"\in (0, R^\mathrm{p})"})
-
-        if self.has_reaction_liquid:
-            symbol_name_ = r"f^{\mathrm{react},\p}_{i}" if self.single_partype else r"f^{\mathrm{react},\p}_{j,i}"
-            dep_ = r"\vec{c}^{\p}, \vec{c}^{\s}; i" if self.single_partype else r"\vec{c}^{\p}, \vec{c}^{\s}; j, i"
-            vars_and_params_.append({"Group" : 10.2, "Symbol": symbol_name_, "Description": r"particle liquid phase reaction function", "Unit": r"\frac{mol}{m^3 \cdot s}", "Dependence": dep_})
-
-        if self.has_reaction_solid:
-            symbol_name_ = r"f^{\mathrm{react},\s}_{i}" if self.single_partype else r"f^{\mathrm{react},\s}_{j,i}"
-            dep_ = r"\vec{c}^{\p}, \vec{c}^{\s}; i" if self.single_partype else r"\vec{c}^{\p}, \vec{c}^{\s}; j, i"
-            vars_and_params_.append({"Group" : 10.3, "Symbol": symbol_name_, "Description": r"particle solid phase reaction function", "Unit": r"\frac{mol}{m^3 \cdot s}", "Dependence": dep_})
-
-        if not self.single_partype:
-            vars_and_params_.append({"Group" : -0.1, "Symbol": r"j", "Description": r"particle type index", "Unit": r"-", "Dependence": r"-", "Property": r""})
-            vars_and_params_.append({"Group" : 1.9, "Symbol": r"d_j", "Description": r"particle type volume fraction", "Unit": r"-", "Dependence": r"particle type", "Property": r""})
-
-        for var_ in vars_and_params_:
-            var_["Symbol"] = format_variables(var_["Symbol"], var_format_)
-        
-        vars_and_params_ = sorted(vars_and_params_, key=lambda x: x['Group'])
-
-        object.__setattr__(self, 'vars_and_params', vars_and_params_)
-
-    
-    def available_CADET_Core(self):
-        return True
-
-    def vars_params_description(self):
-
-        description_ = ""
-
-        idx = 1
-        num_VP = len(self.vars_and_params)
-
-        for thing in self.vars_and_params:
-
-            if thing.get("Group", -1) < 0:
-                num_VP -= 1
-                continue
-
-            if not idx == 1:
-                description_ += ", " if idx < num_VP else ", and "
-                
-            description_ += r"$" + thing["Symbol"]
-
-            if not thing.get("Domain", "-") == "-":
-                description_ += r"\colon " + re.sub(r"\$", "", thing["Domain"]) + r" \mapsto \mathbb{R}"
-
-            description_ += thing.get("Property", "")
-
-            description_ += r"$"
-            
-            description_ += " is the " + thing["Description"]
-
-            idx += 1
-                
-        return description_ + "."
-
-@dataclass
-class Column:
-
-    dev_mode: bool  # dev mode including untested, unstable and wip features
-    advanced_mode: bool  # Ask for detailed parameter inputs
-    resolution: Literal = ""
-    N_c: int = -1
-    N_p: int = -1
-
-    has_axial_coordinate: bool = False
-    has_radial_coordinate: bool = False
-    has_angular_coordinate: bool = False
-    has_axial_dispersion: bool = False
-    has_radial_dispersion: bool = False
-    has_angular_dispersion: bool = False
-
-    # List[bool] = None # TODO make this possible per component and particle type in advanced mode
-    nonlimiting_filmDiff: bool = False
-
-    particle_models: Optional[List[Particle]] = None
-    # counts per unique particle type (geometry, has_core, resolution)
-    par_type_counts: Counter[Particle] = field(default_factory=Counter)
-    # puts particle types together that have a similar contribution to the interstitial volume equation and counts them
-    par_unique_intV_contribution_counts: Counter[Particle] = field(
-        default_factory=Counter)
-    has_surfDiff: bool = False
-    has_binding: bool = True  # and thus solid phase
-    req_binding: bool = False
-    has_mult_bnd_states: bool = False
-    PTD: bool = False
-    # nCompReq: int
-    # nCompKin: int
-    # noPoresButReqBinding: bool
-
-    has_filter: bool = False
-
-    binding_model: str = "Arbitrary"
-
-    has_reaction_bulk: bool = False
-    has_reaction_particle_liquid: bool = False
-    has_reaction_particle_solid: bool = False
-
-    vars_and_params: List[dict] = field(default_factory=list)
-
-    def __post_init__(self):
-
-        # Configure interstitial column transport
-        st.sidebar.write("Configure interstitial volume model")
-        col1, col2 = st.columns(2)
-
-        with col1:
-            self.resolution = re.search(r'\dD', st.sidebar.selectbox("Column resolution", [
-                                        "1D (axial coordinate)", "0D (Homogeneous Tank)", "2D (axial and radial coordinate)", "3D (axial, radial and angular coordinate)"], key=r"column_resolution")).group()
-
-        valid_resolutions = {"3D", "2D", "1D", "0D"}
-
-        if self.resolution not in valid_resolutions:
-            raise ValueError(
-                f"Invalid resolution: {self.resolution}. Must be one of {valid_resolutions}.")
-        if int(re.search("\\d", self.resolution).group()) == 0:
-            self.has_filter = st.sidebar.selectbox(
-                    "Add flow rate filter", ["No", "Yes"], key=r"has_filter") == "Yes"
-        if int(re.search("\\d", self.resolution).group()) > 0:
-            self.has_axial_coordinate = True
-            self.has_axial_dispersion = True
-        if int(re.search("\\d", self.resolution).group()) > 1:
-            self.has_radial_coordinate = True
-            self.has_radial_dispersion = True
-        if int(re.search("\\d", self.resolution).group()) > 2:
-            self.has_angular_coordinate = True
-            self.has_angular_dispersion = True
-
-        self.N_c = st.sidebar.number_input(
-            "Number of components", key=r"N_c", min_value=1, step=1) if dev_mode_ else -1
-
-        if self.has_axial_coordinate:
-            with col2:
-                self.has_axial_dispersion = st.sidebar.selectbox(
-                    "Add axial Dispersion", ["No", "Yes"], key=r"has_axial_dispersion") == "Yes"
-
-        if self.dev_mode:
-            if self.has_radial_coordinate:
-                self.has_radial_dispersion = st.sidebar.selectbox(
-                    "Add radial Dispersion", ["Yes", "No"], key=r"has_radial_dispersion") == "Yes"
-            if self.has_angular_coordinate:
-                self.has_angular_dispersion = st.sidebar.selectbox(
-                    "Add angular Dispersion", ["Yes", "No"], key=r"has_angular_dispersion") == "Yes"
-
-        # Configure particle model
-        st.sidebar.write("Configure particle model")
-
-        self.N_p = int(st.sidebar.selectbox("Add particles", ["No", "Yes"], key=r"add_particles") == "Yes")
-
-        if self.N_p > 0:
-            if dev_mode_:
-                self.N_p = st.sidebar.number_input("Number of particle types", key=r"N^\mathrm{p}", min_value=0, step=1)
-            elif advanced_mode_:
-                self.N_p = 1 + int(st.sidebar.selectbox("Particle size distribution", ["No", "Yes"], key=r"PSD") == "Yes")
-                if self.N_p > 1:
-                    self.PTD = st.sidebar.selectbox("Particle type distribution (binding per j)", ["No", "Yes"], key=r"PTD") == "Yes"
-                
-        if self.N_p > 0:
-
-            self.particle_models = []
-
-            for j in range(self.N_p):
-
-                if self.N_p > 1 and dev_mode_:
-                    st.sidebar.write(f"Configure particle type {j + 1}")
-
-                col1, col2 = st.columns(2)
-
-                with col1:
-
-                    if dev_mode_:  # multiple particle types
-                        resolution = re.search(r'\dD', st.sidebar.selectbox(f"Select spatial resolution of particle type {j + 1}", [
-                                               "1D (radial coordinate)", "0D (homogeneous)"], key=f"parType_{j+1}_resolution")).group()
-                        has_core = st.sidebar.selectbox(f"Choose if particle type {j + 1} is a core-shell particle (i.e. " + r"$R^\mathrm{pc} > 0$)", [
-                                                       "No core-shell", "Has core-shell"], key=f"parType_{j+1}_has_core") == "Has core-shell"
-                        geometry = st.sidebar.selectbox(f"Select geometry of particle type {j + 1}", [
-                                                        "Sphere", "Cylinder", "Slab"], key=f"parType_{j+1}__geometry")
-                    elif j == 0:
-                        resolution = re.search(r'\dD', st.sidebar.selectbox(f"Select spatial resolution of particles", [
-                                               "1D (radial coordinate)", "0D (homogeneous)"], key=r"particle_resolution")).group()
-                        has_core = st.sidebar.selectbox(f"Add impenetrable core-shell to particles (i.e. " + r"$R^\mathrm{pc} > 0$)", [
-                                                       "No", "Yes"], key=f"particle_has_core") == "Yes" if (resolution == "1D" and advanced_mode_) else False
-                        geometry = "Sphere"
-
-                with col2:
-                    if j == 0:  # todo make this configurable for every particle type
-                        self.nonlimiting_filmDiff = st.sidebar.selectbox(
-                            "Non-limiting film diffusion", ["No", "Yes"], key=r"nonlimiting_filmDiff") == "Yes"
-
-                if j == 0:  # todo make this configurable for every particle type
-                    # Configure binding model
-                    st.sidebar.write("Configure binding model")
-
-                    self.has_binding = st.sidebar.selectbox(
-                        "Add binding", ["No", "Yes"], key=r"has_binding") == "Yes"
-
-                    if self.has_binding:
-
-                        self.req_binding = st.sidebar.selectbox("Binding kinetics mode", [
-                                                                "Kinetic", "Rapid-equilibrium"], key=r"req_binding") == "Rapid-equilibrium"
-                        self.binding_model = st.sidebar.selectbox("Binding model", eq.BINDING_MODELS, key=r"binding_model")
-                        self.has_mult_bnd_states = st.sidebar.selectbox("Add multiple bound states", [
-                                                                        "No", "Yes"], key=r"has_mult_bnd_states") == "Yes" if advanced_mode_ else False
-                        self.has_surfDiff = st.sidebar.selectbox("Add surface diffusion", [
-                                                                 "No", "Yes"], key=r"has_surfDiff") == "Yes" if resolution == "1D" else False
-
-                self.particle_models.append(
-                    Particle(
-                        geometry=geometry,
-                        resolution=resolution,
-                        has_core=has_core,
-                        has_binding=self.has_binding,
-                        req_binding=self.req_binding,
-                        has_mult_bnd_states=self.has_mult_bnd_states,
-                        has_surfDiff=self.has_surfDiff,
-                        nonlimiting_filmDiff=self.nonlimiting_filmDiff,
-                        interstitial_volume_resolution=self.resolution,
-                        single_partype=(self.N_p == 1),
-                        PTD=self.PTD,
-                        binding_model=self.binding_model,
-                        has_reaction_liquid=self.has_reaction_particle_liquid,
-                        has_reaction_solid=self.has_reaction_particle_solid
-                    )
-                )
-
-            # We need to count and thus sort particle_models:
-            #  Particle types need individual particle equations, which is why we count them
-            #  Only specific differences lead to changes in the interstitial volume equations: geometry if kinetic film diffusion. else geometry + resolution
-            #  Sorting by (geometry, resolution) ensures we are using the same indices across the interstitial volume and particle eqquations.
-            self.particle_models = sorted(self.particle_models, key=lambda particle: (
-                particle.geometry, particle.resolution))
-            self.par_type_counts = Counter(self.particle_models)
-            if self.nonlimiting_filmDiff:
-                self.par_unique_intV_contribution_counts = Counter(
-                    (particle.geometry, particle.resolution) for particle in self.particle_models)
-            else:
-                self.par_unique_intV_contribution_counts = Counter(
-                    particle.geometry for particle in self.particle_models)
-
-        else:
-            self.has_binding = False
-
-        if advanced_mode_:
-            st.sidebar.write("Configure reactions")
-            self.has_reaction_bulk = st.sidebar.selectbox(
-                "Add bulk liquid reaction", ["No", "Yes"], key=r"has_reaction_bulk") == "Yes"
-            if self.N_p > 0:
-                self.has_reaction_particle_liquid = st.sidebar.selectbox(
-                    "Add particle liquid reaction", ["No", "Yes"], key=r"has_reaction_particle_liquid") == "Yes"
-                self.has_reaction_particle_solid = st.sidebar.selectbox(
-                    "Add particle solid reaction", ["No", "Yes"], key=r"has_reaction_particle_solid") == "Yes"
-
-        self.fill_vars_and_params()
-
-    def available_CADET_Core(self):
-        """
-        Return the availability status of the model in CADET-Core.
-    
-        Returns
-        -------
-        int
-            -1 if model is not present,
-             0 if model can be approximated,
-             1 if model is present.
-        """
-        if self.has_angular_coordinate:
-            return -1
-
-        availability = 1
-
-        par1D = None
-        
-        if self.N_p > 0:
-            
-            par1D = False
-
-            for p in self.particle_models:
-
-                availability = int(availability and p.available_CADET_Core())
-                par1D = par1D or p.resolution == "1D"
-                
-            # no particles with pore diffusion but no film diffusion (yet)
-            if par1D and self.nonlimiting_filmDiff:
-                return 0
-        
-        # tank model
-        if not self.has_axial_coordinate:
-            # only 0D particles with non limiting film diffusion available
-            if self.N_p > 0:
-                if not par1D and self.nonlimiting_filmDiff:
-                    return 1
-                elif par1D and not self.nonlimiting_filmDiff: # can be approximated using 1 FV cell
-                    return 0
-                else:
-                    return -1
-            else:
-                return 1
-
-        return availability
-
-    def available_CADET_Process(self):
-        """
-        Return the availability status of the model in CADET-Process.
-    
-        Returns
-        -------
-        int
-            -1 if model is not present,
-             0 if model can be approximated,
-             1 if model is present.
-        """
-
-        # All CADET-Core models supported except multiple particle type and 2D models
-        
-        core_availability = self.available_CADET_Core()
-        
-        if core_availability == -1:
-            return -1
-        
-        if self.N_p > 1:
-            return -1
-        
-        if self.has_radial_coordinate:
-            return -1
-        
-        else:
-            return core_availability
-
-    def fill_vars_and_params(self):
-
-        without_pores_ = self.nonlimiting_filmDiff and self.has_binding and self.particle_models[0].resolution == "0D"
-
-        state_deps = r"t"
-        param_deps = r"\text{constant}"
-        if self.resolution == "1D":
-            state_deps = r"t, z"
-        if self.resolution == "2D":
-            state_deps = r"t, z, \rho"
-            param_deps = r"\rho"
-        if self.resolution == "3D":
-            state_deps = r"t, z, \rho, \phi"
-            param_deps = r"z, \rho, \phi"
-
-        state_deps += "; i"
-        param_deps_comp = r"i" if param_deps == r"\text{constant}" else param_deps + "; i"
-
-        self.vars_and_params = [
-            {"Group" : 0, "Symbol": r"t", "Description": r"time coordinate", "Unit": r"s", "Dependence": r"\text{independent variable}", "Property": r"\in (0, T^{\mathrm{end}})"},
-            {"Group" : 1, "Symbol": r"c^{\b}_i", "Description": r"bulk liquid concentration", "Unit": r"\frac{mol}{m^3}", "Dependence" : state_deps, "Domain": eq.int_vol_domain(self.resolution)},
-            {"Group" : -1, "Symbol": r"T^{\mathrm{end}}", "Description": r"process end time", "Unit": r"s", "Dependence": r"\text{constant}", "Property": r" > 0"},
-            {"Group" : -0.2, "Symbol": r"N^\mathrm{c}", "Description": r"number of components", "Unit": r"-", "Dependence": r"-", "Property": r"\in \mathbb{N}"},
-            {"Group" : -0.1, "Symbol": r"i", "Description": r"component index", "Unit": r"s", "Dependence": r"-", "Property": r"-"},
-            ]
-
-        if self.has_axial_dispersion:
-            if not without_pores_:
-                self.vars_and_params.append({"Group" : 6, "Symbol": r"D^\mathrm{ax}_i", "Description": r"axial dispersion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": param_deps_comp, "Property": r"\geq 0"})
-            else:
-                self.vars_and_params.append({"Group" : 6, "Symbol": r"\tilde{D}^\mathrm{ax}_i", "Description": r"apparent axial dispersion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": param_deps_comp, "Property": r"\geq 0"})
-        if self.has_radial_dispersion:
-            self.vars_and_params.append({"Group" : 6, "Symbol": r"D^\mathrm{rad}_i", "Description": r"radial dispersion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": param_deps_comp, "Property": r"\geq 0"})
-        if self.has_angular_dispersion:
-            self.vars_and_params.append({"Group" : 6, "Symbol": r"D^\mathrm{ang}_i", "Description": r"angular dispersion coefficient", "Unit": r"\frac{m^2}{s}", "Dependence": param_deps_comp, "Property": r"\geq 0"})
-
-        if self.resolution == "0D":
-            self.vars_and_params.append({"Group" : 2, "Symbol": r"Q^\mathrm{in}", "Description": r"volumetric flow rate into the tank", "Unit": r"\frac{m^3}{s}", "Dependence": r"\text{constant}", "Property": r"\geq 0"})
-            self.vars_and_params.append({"Group" : 2, "Symbol": r"Q^\mathrm{out}", "Description": r"volumetric flow rate out of the tank", "Unit": r"\frac{m^3}{s}", "Dependence": r"\text{constant}", "Property": r"\geq 0"})
-            if self.has_filter:
-                self.vars_and_params.append({"Group" : 2, "Symbol": r"Q^\mathrm{filter}", "Description": r"volumetric flow rate out of the tank (solvent only)", "Unit": r"\frac{m^3}{s}", "Dependence": r"\text{constant}", "Property": r"\geq 0"})
-        else:
-            self.vars_and_params.append({"Group" : 0, "Symbol": r"z", "Description": r"axial cylinder coordinate", "Unit": r"m", "Dependence": r"\text{independent variable}", "Property": r"\in (0, L)"})
-            self.vars_and_params.append({"Group" : -1, "Symbol": r"L", "Description": r"length of cylinder", "Unit": r"m", "Dependence": r"\text{constant}", "Property": r" > 0"})
-            if not without_pores_:
-                self.vars_and_params.append({"Group" : 5, "Symbol": r"u", "Description": r"interstitial velocity", "Unit": r"\frac{m}{s}", "Dependence": param_deps, "Property": r"> 0"})
-            else:
-                self.vars_and_params.append({"Group" : 5, "Symbol": r"\tilde{u}", "Description": r"apparent interstitial velocity", "Unit": r"\frac{m}{s}", "Dependence": param_deps, "Property": r"> 0"})
-
-        if self.resolution == "2D":
-            self.vars_and_params.append({"Group" : 0, "Symbol": r"\rho", "Description": r"radial cylinder coordinate", "Unit": r"m", "Dependence": r"\text{independent variable}", "Property": r"\in (0, R^{\mathrm{c}})"})
-            self.vars_and_params.append({"Group" : -1, "Symbol": r"R^{\mathrm{c}}", "Description": r"cylinder radius", "Unit": r"m", "Dependence": r"\text{constant}", "Property": r" > 0"})
-
-        if self.resolution == "3D":
-            self.vars_and_params.append({"Group" : 0, "Symbol": r"\phi", "Description": r"angular cylinder coordinate", "Unit": r"m", "Dependence": r"\text{independent variable}", "Property": r"\in (0, 2\pi)"})
-
-        if self.N_p > 0:
-            if not without_pores_:
-                self.vars_and_params.append({"Group" : 0.1, "Symbol": r"R^\mathrm{p}", "Description": r"particle radius", "Unit": r"-", "Dependence": r"-", "Property": r"> 0"})
-            if not without_pores_:
-                self.vars_and_params.append({"Group" : 4, "Symbol": r"\varepsilon^{\mathrm{c}}", "Description": r"column porosity", "Unit": r"-", "Dependence": re.sub("t, ", "", state_deps), "Property": r"\in (0, 1)"})
-            else:
-                self.vars_and_params.append({"Group" : 4, "Symbol": r"\varepsilon^{\mathrm{t}}", "Description": r"total porosity", "Unit": r"-", "Dependence": re.sub("t, ", "", state_deps), "Property": r"\in (0, 1)"})
-            if not self.nonlimiting_filmDiff:
-                symbol_name_ = r"k^\mathrm{f}_{i}" if self.N_p<=1 else r"k^\mathrm{f}_{j,i}"
-                self.vars_and_params.append({"Group" : 7, "Symbol": symbol_name_, "Description": r"film diffusion coefficient", "Unit": r"\frac{m}{s}", "Dependence": r"i" if self.N_p<=1 else r"j,i", "Property": r"> 0"})
-
-        if self.has_reaction_bulk:
-            self.vars_and_params.append({"Group" : 8, "Symbol": r"f^{\mathrm{react},\b}_{i}", "Description": r"bulk liquid phase reaction function", "Unit": r"\frac{mol}{m^3 \cdot s}", "Dependence": r"\vec{c}^{\b}; i"})
-            self.vars_and_params.append({"Group" : 8.1, "Symbol": r"\vec{c}^{\b}", "Description": r"bulk liquid components vector", "Unit": r"[\frac{mol}{m^3}]", "Dependence": state_deps})
-
-        for var_ in self.vars_and_params:
-            var_["Symbol"] = format_variables(var_["Symbol"], var_format_)
-            
-        self.vars_and_params = sorted(self.vars_and_params, key=lambda x: x['Group'])
-
-    def interstitial_volume_equation(self):
-
-        without_pores_ = self.nonlimiting_filmDiff and self.has_binding and self.particle_models[0].resolution == "0D"       
-
-        if self.resolution == "0D":
-
-            filter_str = r" - Q_{\mathrm{filter}}" if self.has_filter else ""
-
-            equation = r"""
-    \frac{\mathrm{d}V^{\b}}{\mathrm{d}t} &= Q_{\mathrm{in}} - Q_{\mathrm{out}}""" + filter_str + r""",
-    \\
-    \frac{\mathrm{d}}{\mathrm{d} t} \left( V^{\b} c^{\b}_i \right)"""
-
-            if self.nonlimiting_filmDiff and self.has_binding and self.particle_models[0].resolution == "0D":
-                equation += r" + V^{\p} \varepsilon^{\mathrm{p}} \frac{\partial c^{b}_i}{\partial t}"
-                if self.req_binding:
-                    equation += r" + V^{\p} \left( 1 - \varepsilon^{\mathrm{p}} \right) \frac{\partial c^{\s}_i}{\partial t}"
-
-            equation += r"&= Q_{\mathrm{in}} c^{\b}_{\mathrm{in},i} - Q_{\mathrm{out}} c^{\b}_i"
-
-            if self.nonlimiting_filmDiff and self.has_binding and self.particle_models[0].resolution == "0D" and not self.req_binding:
-                equation += r" - V^{\p} \left( 1 - \varepsilon^{\mathrm{p}} \right) \frac{\partial c^{\s}_i}{\partial t}"
-
-            if self.has_reaction_bulk:
-                equation += " + " + eq.bulk_reaction_term()
-
-        else:
-            equation = eq.bulk_time_derivative(r"\varepsilon^{\mathrm{c}}") if not without_pores_ else eq.bulk_time_derivative()
-            if without_pores_:
-                equation += r" + " + eq.solid_time_derivative(r"\varepsilon^{\mathrm{t}}")
-
-            needs_linebreak = self.has_radial_dispersion or self.has_angular_dispersion
-            eq_sign = " &= " if needs_linebreak else " = "
-
-            equation += eq_sign + eq.axial_convection() if without_pores_ else eq_sign + eq.axial_convection(r"\varepsilon^{\mathrm{c}}")
-
-            if self.has_axial_dispersion:
-                equation += " + " + eq.axial_dispersion(r"\varepsilon^{\mathrm{c}}")
-            if self.has_radial_dispersion:
-                equation += r" \nonumber \\ & + " + eq.radial_dispersion(r"\varepsilon^{\mathrm{c}}")
-            if self.has_angular_dispersion:
-                equation += r" \nonumber \\ & + " + eq.angular_dispersion(r"\varepsilon^{\mathrm{c}}")
-
-            if self.N_p == 0:  # remove occurencies of porosity, which is just constant one in this case
-                equation = re.sub(r"\\varepsilon^{\\mathrm{c}}", "", re.sub(
-                    r"\\left\( \\varepsilon^{\\mathrm{c}} c^{\\l}_i \\right\)", r"c^{\\l}_i", equation))
-
-        # if self.nonlimiting_filmDiff and 1Dparticle # entscheidende faktoren sind particle resolution und filmDiffMode. the following loop has thus to change
-        par_added = 0
-        for par_uniq in self.par_unique_intV_contribution_counts.keys():
-
-            if dev_mode_:
-                equation += eq.int_filmDiff_term(
-                    Particle(
-                        self.particle_models[par_added].geometry, self.particle_models[par_added].has_core, self.particle_models[par_added].resolution
-                    ),
-                    1 + par_added, par_added + self.par_unique_intV_contribution_counts[par_uniq], self.N_p == 1, self.nonlimiting_filmDiff, self.particle_models[par_added].has_surfDiff
-                )
-            else:
-                equation += eq.int_filmDiff_term(
-                    Particle(
-                        self.particle_models[0].geometry, self.particle_models[0].has_core, self.particle_models[0].resolution
-                    ),
-                    1, r"N^{\mathrm{p}}", self.N_p == 1, self.nonlimiting_filmDiff, self.particle_models[0].has_surfDiff
-                )
-
-            par_added += self.par_unique_intV_contribution_counts[par_uniq]
-
-        if self.has_reaction_bulk and self.resolution != "0D":
-            equation += " + " + eq.bulk_reaction_term()
-
-        if self.resolution == "0D":
-            equation = re.sub(
-                r"\\left\(1 - \\varepsilon^{\\mathrm{c}} \\right\)", r"V^s", equation)
-
-        equation = r"""\begin{align}
-""" + equation + r""",
-\end{align}"""
-
-        return equation
-
-    def interstitial_volume_bc(self):
-        if not self.resolution == "0D":
-            return eq.int_vol_BC(self.resolution, self.has_axial_dispersion)
-        else:
-            return None
-
-    def particle_equations(self):
-
-        eqs = {}
-        boundary_conditions = {}
-
-        for par_type in self.par_type_counts.keys():
-
-            eqs[par_type] = eq.particle_transport(par_type, singleParticle=self.N_p == 1, nonlimiting_filmDiff=self.nonlimiting_filmDiff,
-                                                  has_surfDiff=self.has_surfDiff, has_binding=self.has_binding, req_binding=self.req_binding, has_mult_bnd_states=self.has_mult_bnd_states, PTD=self.PTD,
-                                                  has_reaction_liquid=self.has_reaction_particle_liquid, has_reaction_solid=self.has_reaction_particle_solid,
-                                                  binding_model=self.binding_model)
-            eqs[par_type] = eqs[par_type]
-
-            boundary_conditions[par_type] = eq.particle_boundary(par_type, singleParticle=self.N_p == 1, nonlimiting_filmDiff=self.nonlimiting_filmDiff,
-                                                                 has_surfDiff=self.has_surfDiff, has_binding=self.has_binding, req_binding=self.req_binding, has_mult_bnd_states=self.has_mult_bnd_states)
-            boundary_conditions[par_type] = boundary_conditions[par_type]
-
-            if self.N_p == 1:
-                eqs[par_type] = re.sub(",j", "", eqs[par_type])
-                eqs[par_type] = re.sub("j,", "", eqs[par_type])
-                eqs[par_type] = re.sub("_{j}", "", eqs[par_type])
-                boundary_conditions[par_type] = re.sub(",j", "", boundary_conditions[par_type])
-                boundary_conditions[par_type] = re.sub("j,", "", boundary_conditions[par_type])
-                boundary_conditions[par_type] = re.sub("_{j}", "", boundary_conditions[par_type])
-
-        return eqs, boundary_conditions
-
-    def domain_interstitial(self, with_time_domain=True):
-        return r"$" + eq.int_vol_domain(self.resolution, with_time_domain=with_time_domain) + r"$"
-
-    def domain_particle(self):
-        if self.N_p > 0:
-            return eq.full_particle_conc_domain(self.resolution, self.particle_models[0].resolution, self.particle_models[0].has_core, False, False)
-        else:
-            return ""
-        
-    def model_name(self):
-
-        if self.resolution == "0D":
-
-            if self.N_p > 0:
-                model_name = "Finite Bath"
-
-                if self.particle_models[0].resolution == "0D":
-                    model_name += " without Pores" if self.nonlimiting_filmDiff else " with Pores"
-                else:
-                    model_name = "General " + model_name
-            else:
-                return "Continuously Stirred Tank"
-
-        else:
-
-            if self.has_angular_coordinate:
-                model_name = "3D "
-            elif self.has_radial_coordinate:
-                model_name = "2D "
-            # elif self.has_axial_coordinate: # default case, no name prefix !
-            else:
-                model_name = ""
-
-            if self.N_p > 0:
-
-                if self.particle_models[0].resolution == "1D":
-                    model_name += "General Rate Model"
-
-                else:
-                    model_name += "Lumped Rate Model"
-
-                    model_name += " without Pores" if self.nonlimiting_filmDiff else " with Pores"
-
-                if self.N_p > 1:
-                    if not any(par.geometry != "Sphere" for par in self.particle_models):
-                        model_name += " with particle-size distribution"  # particle-size distribution
-                    else:
-                        # particle-type distribution # TODO use when different kinds of geometry or binding
-                        model_name += " with particle-type distribution"
-            else:
-                if self.has_axial_dispersion or self.has_radial_dispersion or self.has_angular_dispersion:
-                    model_name += "Dispersive "
-                model_name += "Plug Flow"  # Reactor if we have reactions
-
-        if self.has_binding and self.binding_model != "Arbitrary":
-            model_name += f" with {self.binding_model} binding"
-
-        return model_name
-
-    def model_assumptions(self):
-
-        asmpts = {
-            "General model assumptions": eq.HRM_asmpt(self.N_p, self.nonlimiting_filmDiff, self.has_binding, self.has_surfDiff, self.resolution),
-            "Specific model assumptions": eq.int_vol_continuum_asmpt(self.resolution, self.N_p, self.nonlimiting_filmDiff) +
-            (eq.particle_asmpt(
-                self.particle_models[0].resolution, self.has_surfDiff) if self.N_p > 0 else [])
-        }
-
-        if self.nonlimiting_filmDiff:
-            asmpts["Specific model assumptions"].append(
-                r"the film around the particles does not limit mass transfer. That is, we assume $k^{\mathrm{f}}_i = \infty$)")
-        if self.req_binding:
-            asmpts["Specific model assumptions"].append(
-                r"adsorption and desorption happen on a much faster time scale than the other mass transfer processes (e.g., convection, diffusion). Hence, we consider them to be equilibrated instantly, that is, to always be in (local) equilibrium")
-
-        for idx in range(0, len(asmpts["Specific model assumptions"])):
-            
-            asmpts["Specific model assumptions"][idx] = format_variables(asmpts["Specific model assumptions"][idx], var_format_)
-            
-        for idx in range(0, len(asmpts["General model assumptions"])):
-            
-            asmpts["General model assumptions"][idx] = format_variables(asmpts["General model assumptions"][idx], var_format_)
-            
-        if not self.binding_model == "Arbitrary":
-            
-            asmpts.update({"Binding model assumptions": eq.binding_model_assumptions(self.binding_model)})
-            
-        return asmpts
-
-    def vars_params_description(self):
-
-        description_ = ""
-
-        idx_ = 1
-        num_VP = len(self.vars_and_params)
-
-        for thing in self.vars_and_params:
-
-            if thing.get("Group", -1) < 0: # dont print symbols with negative group no.
-                num_VP -= 1
-                continue
-
-            if not idx_ == 1:
-                description_ += ", " if idx_ < num_VP else ", and "
-            description_ += r"$" + thing["Symbol"]
-
-            if not thing.get("Domain", "-") == "-":
-                description_ += r"\colon " + re.sub(r"\$", "", thing["Domain"]) + r" \mapsto \mathbb{R}"
-
-            description_ += thing.get("Property", "") + r"$"
-            
-            description_ += " is the " + thing["Description"]
-
-            idx_ += 1
-                
-        return description_ + "."
+from src.model_column import Column
 
 # %% Streamlit UI
 
@@ -855,7 +115,7 @@ if advanced_mode_:
 else:
     dev_mode_ = False
 
-column_model = Column(dev_mode=dev_mode_, advanced_mode=advanced_mode_)
+column_model = Column(dev_mode=dev_mode_, advanced_mode=advanced_mode_, var_format=var_format_)
 
 # %% Display equations
